requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==23.0.0
drf-nested-routers==0.93.4
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
dj-database-url  # Para manejar la configuración de la base de datos con variables de entorno